+++ tests/test_api_requests.py
@@ -147,7 +147,7 @@ UNSET = object()
     ],
 )
 async def test_enabled(setting, enabled, mockserver):
-    settings: SETTINGS_T = {}
+    settings = {}
     if setting is not UNSET:
         settings["ZYTE_API_ENABLED"] = setting
     async with mockserver.make_handler(settings) as handler:
@@ -162,7 +162,7 @@ async def test_enabled(setting, enabled, mockserver):
 async def test_coro_handling(zyte_api: bool, mockserver):
     """ScrapyZyteAPIDownloadHandler.download_request must return a deferred
     both when using Zyte API and when using the regular downloader logic."""
-    settings: SETTINGS_T = {"ZYTE_API_DEFAULT_PARAMS": {"browserHtml": True}}
+    settings = {"ZYTE_API_DEFAULT_PARAMS": {"browserHtml": True}}
     async with mockserver.make_handler(settings) as handler:
         req = Request(
             # this should really be a URL to a website, not to the API server,
@@ -258,7 +258,6 @@ async def test_higher_concurrency():
 
         crawler = get_crawler(
             {
-                **SETTINGS,
                 "CONCURRENT_REQUESTS": concurrency,
                 "CONCURRENT_REQUESTS_PER_DOMAIN": concurrency,
                 "ZYTE_API_URL": server.urljoin("/"),
@@ -304,7 +303,7 @@ async def test_params_parser_input_default(mockserver):
 
 @ensureDeferred
 async def test_param_parser_input_custom(mockserver):
-    settings: SETTINGS_T = {
+    settings = {
         "ZYTE_API_EXPERIMENTAL_COOKIES_ENABLED": True,
         "ZYTE_API_AUTOMAP_PARAMS": {"c": "d"},
         "ZYTE_API_BROWSER_HEADERS": {"B": "b"},
@@ -442,7 +441,7 @@ def test_transparent_mode_toggling(setting, meta, expected):
     :func:`~scrapy_zyte_api.handler._get_api_params` parameter.
     """
     request = Request(url="https://example.com", meta=meta)
-    settings: SETTINGS_T = {**SETTINGS, "ZYTE_API_TRANSPARENT_MODE": setting}
+    settings = {"ZYTE_API_TRANSPARENT_MODE": setting}
     crawler = get_crawler(settings)
     handler = get_download_handler(crawler, "https")
     param_parser = handler._param_parser
@@ -497,7 +496,7 @@ async def test_job_id(meta, mockserver):
     """
     request = Request(url="https://example.com", meta={meta: True})
     with set_env(SHUB_JOBKEY="1/2/3"):
-        crawler = get_crawler(SETTINGS)
+        crawler = get_crawler()
         handler = get_download_handler(crawler, "https")
         param_parser = handler._param_parser
         api_params = param_parser.parse(request)
@@ -518,7 +517,7 @@ async def test_default_params_none(mockserver, caplog):
     the settings are mapped to the corresponding
     :func:`~scrapy_zyte_api.handler._get_api_params` parameter.
     """
-    settings: SETTINGS_T = {
+    settings = {
         "ZYTE_API_DEFAULT_PARAMS": {"a": None, "b": "c"},
         "ZYTE_API_AUTOMAP_PARAMS": {"d": None, "e": "f"},
     }
@@ -561,7 +560,7 @@ async def test_default_params_none(mockserver, caplog):
         (
             "ZYTE_API_AUTOMAP_PARAMS",
             "zyte_api_automap",
-            {"httpResponseBody", "httpResponseHeaders"},
+            set(DEFAULT_AUTOMAP_PARAMS),
         ),
     ],
 )
@@ -582,8 +581,7 @@ def test_default_params_merging(
     """
     request = Request(url="https://example.com")
     request.meta[meta_key] = meta
-    settings: SETTINGS_T = {**SETTINGS, setting_key: setting}
-    crawler = get_crawler(settings)
+    crawler = get_crawler({setting_key: setting})
     handler = get_download_handler(crawler, "https")
     param_parser = handler._param_parser
     with caplog.at_level("WARNING"):
@@ -636,8 +634,7 @@ def test_default_params_immutability(setting_key, meta_key, setting, meta):
     request = Request(url="https://example.com")
     request.meta[meta_key] = meta
     default_params = copy(setting)
-    settings: SETTINGS_T = {**SETTINGS, setting_key: setting}
-    crawler = get_crawler(settings)
+    crawler = get_crawler({setting_key: setting})
     handler = get_download_handler(crawler, "https")
     param_parser = handler._param_parser
     param_parser.parse(request)
@@ -646,7 +643,7 @@ def test_default_params_immutability(setting_key, meta_key, setting, meta):
 
 @inlineCallbacks
 def _test_automap(
-    settings: SETTINGS_T,
+    settings,
     request_kwargs,
     meta,
     expected,
@@ -656,7 +653,7 @@ def _test_automap(
 ):
     request = Request(url="https://example.com", **request_kwargs)
     request.meta["zyte_api_automap"] = meta
-    settings = {**SETTINGS, **settings, "ZYTE_API_TRANSPARENT_MODE": True}
+    settings = {**settings, "ZYTE_API_TRANSPARENT_MODE": True}
     crawler = get_crawler(settings)
     yield _process_request(crawler, request, is_start_request=True)
     if "cookies" in request_kwargs:
@@ -768,6 +765,18 @@ def _test_automap(
                 "httpResponseBody": True,
                 "httpResponseHeaders": True,
             },
+            [],
+        ),
+        # To request httpResponseHeaders on their own, you must disable
+        # httpResponseBody.
+        (
+            {"httpResponseHeaders": True},
+            {"httpResponseBody": True, "httpResponseHeaders": True},
+            [],
+        ),
+        (
+            {"httpResponseBody": False, "httpResponseHeaders": True},
+            {"httpResponseHeaders": True},
             [],
         ),
     ],
@@ -1570,6 +1579,456 @@ def test_automap_method(method, meta, expected, warnings, caplog):
             },
             [],
         ),
+        # Zyte Smart Proxy Manager special header handling.
+        (
+            {"X-Crawlera-Foo": "Bar"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Client": "Custom client string"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "enable"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["To achieve the same behavior with Zyte API, do not set request cookies"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "disable"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["it is the default behavior of Zyte API"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "discard"},
+            {},
+            {
+                "cookieManagement": "discard",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "foo"},
+            {
+                "cookieManagement": "bar",
+            },
+            {
+                "cookieManagement": "bar",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "foo"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["cannot be mapped to a Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-JobId": "foo"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+                "jobId": "foo",
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-JobId": "foo"},
+            {
+                "jobId": "bar",
+            },
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+                "jobId": "bar",
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Max-Retries": "1"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-No-Bancheck": "1"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "pass"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["cannot be mapped to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Profile": "desktop"},
+            {},
+            {
+                "device": "desktop",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Profile": "mobile"},
+            {},
+            {
+                "device": "mobile",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Profile": "foo"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["cannot be mapped to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Profile": "foo"},
+            {
+                "device": "bar",
+            },
+            {
+                "device": "bar",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Profile-Pass": "foo"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Region": "foo"},
+            {},
+            {
+                "geolocation": "foo",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Region": "foo"},
+            {
+                "geolocation": "bar",
+            },
+            {
+                "geolocation": "bar",
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Session": "foo"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Timeout": "40000"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Use-Https": "1"},
+            {},
+            {
+                "httpResponseBody": True,
+                "httpResponseHeaders": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Foo": "Bar"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Client": "Custom client string"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "enable"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["To achieve the same behavior with Zyte API, do not set request cookies"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "disable"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["it is the default behavior of Zyte API"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "discard"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+                "cookieManagement": "discard",
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "foo"},
+            {
+                "browserHtml": True,
+                "cookieManagement": "bar",
+            },
+            {
+                "browserHtml": True,
+                "cookieManagement": "bar",
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Cookies": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["cannot be mapped to a Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-JobId": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+                "jobId": "foo",
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-JobId": "foo"},
+            {
+                "browserHtml": True,
+                "jobId": "bar",
+            },
+            {
+                "browserHtml": True,
+                "jobId": "bar",
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Max-Retries": "1"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-No-Bancheck": "1"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "pass"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "desktop"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "mobile"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile": "foo"},
+            {
+                # Zyte API does not support it, it will trigger a 400 response,
+                # but we allow it for forward compatibility, i.e. in case it is
+                # supported in the future.
+                "device": "bar",
+                "browserHtml": True,
+            },
+            {
+                "device": "bar",
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Profile-Pass": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Region": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+                "geolocation": "foo",
+            },
+            ["has been assigned to the matching Zyte API request parameter"],
+        ),
+        (
+            {"X-Crawlera-Region": "foo"},
+            {
+                "browserHtml": True,
+                "geolocation": "bar",
+            },
+            {
+                "browserHtml": True,
+                "geolocation": "bar",
+            },
+            ["has already been defined on the request"],
+        ),
+        (
+            {"X-Crawlera-Session": "foo"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Timeout": "40000"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
+        (
+            {"X-Crawlera-Use-Https": "1"},
+            {
+                "browserHtml": True,
+            },
+            {
+                "browserHtml": True,
+            },
+            ["This header has been dropped"],
+        ),
     ],
 )
 @inlineCallbacks
@@ -2146,8 +2605,7 @@ def test_automap_all_cookies(meta):
     """Because of scenarios like cross-domain redirects and browser rendering,
     Zyte API requests should include all cookie jar cookies, regardless of
     the target URL domain."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings: Dict[str, Any] = {
         "ZYTE_API_EXPERIMENTAL_COOKIES_ENABLED": True,
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
@@ -2254,8 +2712,7 @@ def test_automap_cookie_jar(meta):
         url="https://example.com/3", meta={**meta, "cookiejar": "a"}, cookies={"x": "w"}
     )
     request4 = Request(url="https://example.com/4", meta={**meta, "cookiejar": "a"})
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings: Dict[str, Any] = {
         "ZYTE_API_EXPERIMENTAL_COOKIES_ENABLED": True,
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
@@ -2306,8 +2763,7 @@ def test_automap_cookie_jar(meta):
     ],
 )
 def test_automap_cookie_limit(meta, caplog):
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings: Dict[str, Any] = {
         "ZYTE_API_EXPERIMENTAL_COOKIES_ENABLED": True,
         "ZYTE_API_MAX_COOKIES": 1,
         "ZYTE_API_TRANSPARENT_MODE": True,
@@ -2436,8 +2892,7 @@ class CustomCookieMiddleware(CookiesMiddleware):
 
 def test_automap_custom_cookie_middleware():
     mw_cls = CustomCookieMiddleware
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "DOWNLOADER_MIDDLEWARES": {
             "scrapy.downloadermiddlewares.cookies.CookiesMiddleware": None,
             f"{mw_cls.__module__}.{mw_cls.__qualname__}": 700,
@@ -2635,8 +3090,7 @@ def test_default_params_automap(default_params, meta, expected, warnings, caplog
     ``ZYTE_API_AUTOMAP_PARAMS`` setting."""
     request = Request(url="https://example.com")
     request.meta["zyte_api_automap"] = meta
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_AUTOMAP_PARAMS": default_params,
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
@@ -2665,8 +3119,7 @@ def test_default_params_false(default_params):
     """If zyte_api_default_params=False is passed, ZYTE_API_DEFAULT_PARAMS is ignored."""
     request = Request(url="https://example.com")
     request.meta["zyte_api_default_params"] = False
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_DEFAULT_PARAMS": default_params,
     }
     crawler = get_crawler(settings)
@@ -2706,11 +3159,7 @@ def _process_request(crawler, request, is_start_request=False):
 def test_middleware_headers_start_requests():
     """By default, automap should not generate a customHttpRequestHeaders
     parameter."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
-        "ZYTE_API_TRANSPARENT_MODE": True,
-    }
-    crawler = get_crawler(settings)
+    crawler = get_crawler({"ZYTE_API_TRANSPARENT_MODE": True})
     request = Request(url="https://example.com")
     yield _process_request(crawler, request, is_start_request=True)
 
@@ -2724,11 +3173,7 @@ def test_middleware_headers_start_requests():
 def test_middleware_headers_cb_requests():
     """Callback requests will include the Referer parameter if the Referer
     middleware is not disabled."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
-        "ZYTE_API_TRANSPARENT_MODE": True,
-    }
-    crawler = get_crawler(settings)
+    crawler = get_crawler({"ZYTE_API_TRANSPARENT_MODE": True})
     request = Request(url="https://example.com")
     yield _process_request(crawler, request)
 
@@ -2744,8 +3189,7 @@ def test_middleware_headers_cb_requests():
 def test_middleware_headers_cb_requests_skip():
     """Callback requests will not include the Referer parameter if the Referer
     header is configured to be skipped."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_SKIP_HEADERS": list(
             set(header.decode() for header in SKIP_HEADERS)
             | {
@@ -2770,8 +3214,7 @@ def test_middleware_headers_default():
     the global default, and values matching defaults from middlewares that are
     ignored otherwise, its headers should be translated into the
     customHttpRequestHeaders parameter."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "DEFAULT_REQUEST_HEADERS": {
             **DEFAULT_REQUEST_HEADERS,
             "Accept-Encoding": DEFAULT_ACCEPT_ENCODING,
@@ -2808,8 +3251,7 @@ def test_middleware_headers_default():
 def test_middleware_headers_default_custom():
     """Non-default values set for headers with a default value also work as
     expected."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "DEFAULT_REQUEST_HEADERS": {
             "Accept": "text/html",
             "Accept-Language": "fa",
@@ -2846,8 +3288,7 @@ def test_middleware_headers_default_custom():
 def test_middleware_headers_default_skip():
     """Headers set through DEFAULT_REQUEST_HEADERS will not be translated into
     the customHttpRequestHeaders parameter if configured to be skipped."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "DEFAULT_REQUEST_HEADERS": {
             **DEFAULT_REQUEST_HEADERS,
             "Accept-Encoding": DEFAULT_ACCEPT_ENCODING,
@@ -2873,8 +3314,7 @@ def test_middleware_headers_default_skip():
 def test_middleware_headers_request_headers():
     """If request headers match the global default value of
     DEFAULT_REQUEST_HEADERS, they should be translated nonetheless."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
     crawler = get_crawler(settings)
@@ -2910,11 +3350,7 @@ def test_middleware_headers_request_headers():
 def test_middleware_headers_request_headers_custom():
     """Non-default values set for headers with a default value also work as
     expected."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
-        "ZYTE_API_TRANSPARENT_MODE": True,
-    }
-    crawler = get_crawler(settings)
+    crawler = get_crawler({"ZYTE_API_TRANSPARENT_MODE": True})
     request = Request(
         url="https://example.com",
         headers={
@@ -2949,8 +3385,7 @@ def test_middleware_headers_request_headers_custom():
 def test_middleware_headers_request_headers_skip():
     """Headers set on the request will not be translated into the
     customHttpRequestHeaders parameter if configured to be skipped."""
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_SKIP_HEADERS": list(
             set(header.decode() for header in SKIP_HEADERS)
             | {*DEFAULT_REQUEST_HEADERS, "Accept-Encoding", "Referer", "User-Agent"}
@@ -2991,14 +3426,14 @@ def test_middleware_headers_custom_middleware_before():
     DEFAULT_REQUEST_HEADERS, they will *not* be translated."""
 
     settings: SETTINGS_T = {
-        **SETTINGS,
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
     mw1 = "tests.test_api_requests.DefaultValuesDownloaderMiddleware"
     mw2 = "scrapy_zyte_api.ScrapyZyteAPIDownloaderMiddleware"
-    settings["DOWNLOADER_MIDDLEWARES"][mw1] = (
-        settings["DOWNLOADER_MIDDLEWARES"][mw2] - 1
-    )
+    settings["DOWNLOADER_MIDDLEWARES"] = {
+        **SETTINGS["DOWNLOADER_MIDDLEWARES"],
+        mw1: SETTINGS["DOWNLOADER_MIDDLEWARES"][mw2] - 1,
+    }
     crawler = get_crawler(settings)
     request = Request("https://example.com")
     yield _process_request(crawler, request)
@@ -3029,14 +3464,14 @@ def test_middleware_headers_custom_middleware_before_custom():
     the scrapy-zyte-api downloader middleware have non-default values, they
     will be translated."""
     settings: SETTINGS_T = {
-        **SETTINGS,
         "ZYTE_API_TRANSPARENT_MODE": True,
     }
     mw1 = "tests.test_api_requests.CustomValuesDownloaderMiddleware"
     mw2 = "scrapy_zyte_api.ScrapyZyteAPIDownloaderMiddleware"
-    settings["DOWNLOADER_MIDDLEWARES"][mw1] = (
-        settings["DOWNLOADER_MIDDLEWARES"][mw2] - 1
-    )
+    settings["DOWNLOADER_MIDDLEWARES"] = {
+        **SETTINGS["DOWNLOADER_MIDDLEWARES"],
+        mw1: SETTINGS["DOWNLOADER_MIDDLEWARES"][mw2] - 1,
+    }
     crawler = get_crawler(settings)
     request = Request("https://example.com")
     yield _process_request(crawler, request)
@@ -3068,8 +3503,7 @@ def test_middleware_headers_custom_middleware_before_skip():
     the scrapy-zyte-api downloader middleware will not be translated into the
     customHttpRequestHeaders parameter if configured to be skipped."""
 
-    settings: SETTINGS_T = {
-        **SETTINGS,
+    settings = {
         "ZYTE_API_SKIP_HEADERS": list(
             set(header.decode() for header in SKIP_HEADERS)
             | {*DEFAULT_REQUEST_HEADERS, "Accept-Encoding", "Referer", "User-Agent"}
@@ -3078,9 +3512,10 @@ def test_middleware_headers_custom_middleware_before_skip():
     }
     mw1 = "tests.test_api_requests.CustomValuesDownloaderMiddleware"
     mw2 = "scrapy_zyte_api.ScrapyZyteAPIDownloaderMiddleware"
-    settings["DOWNLOADER_MIDDLEWARES"][mw1] = (
-        settings["DOWNLOADER_MIDDLEWARES"][mw2] - 1
-    )
+    settings["DOWNLOADER_MIDDLEWARES"] = {
+        **SETTINGS["DOWNLOADER_MIDDLEWARES"],
+        mw1: SETTINGS["DOWNLOADER_MIDDLEWARES"][mw2] - 1,
+    }
     crawler = get_crawler(settings)
     request = Request("https://example.com")
     yield _process_request(crawler, request)